+++ internal/tui/dashboard.go
@@ -267,9 +267,8 @@ func renderMemContent(mem collector.MemSnapshot, innerW int) string {
 		sb.WriteString(renderGauge(mem.SwapPercent, gaugeW))
 		sb.WriteString(valueStyle.Render(fmt.Sprintf(" %5.1f", mem.SwapPercent)))
 		sb.WriteString(unitStyle.Render("%\n"))
-		sb.WriteString(dimStyle.Render(fmt.Sprintf("  %s / %s\n",
-			formatBytes(mem.SwapUsedBytes), formatBytes(mem.SwapTotalBytes))))
-		sb.WriteString(dimStyle.Render(fmt.Sprintf("  in %s  out %s\n",
+		sb.WriteString(dimStyle.Render(fmt.Sprintf("  %s / %s  in %s  out %s\n",
+			formatBytes(mem.SwapUsedBytes), formatBytes(mem.SwapTotalBytes),
 			formatRate(mem.SwapInRate), formatRate(mem.SwapOutRate))))
 	}
 
@@ -301,13 +300,10 @@ func renderDiskContent(disks []collector.DiskSnapshot, io []collector.DiskIOSnap
 				ioSnap = ioByDevice[parent]
 			}
 		}
-		sb.WriteString(dimStyle.Render(fmt.Sprintf("  %s / %s  r %s  w %s",
+		sb.WriteString(dimStyle.Render(fmt.Sprintf("  %s / %s  r %s  w %s  util %.0f%%  await %.1fms",
 			formatBytes(d.UsedBytes), formatBytes(d.TotalBytes),
-			formatRate(ioSnap.ReadRate), formatRate(ioSnap.WriteRate))))
-		sb.WriteString("\n")
-		sb.WriteString(dimStyle.Render(fmt.Sprintf("   util %4.0f%%", ioSnap.UtilPercent)))
-		sb.WriteString("\n")
-		sb.WriteString(dimStyle.Render(fmt.Sprintf("   await %5.1fms", ioSnap.LatencyMs)))
+			formatRate(ioSnap.ReadRate), formatRate(ioSnap.WriteRate),
+			ioSnap.UtilPercent, ioSnap.LatencyMs)))
 		sb.WriteString("\n")
 	}
 	return sb.String()